Many changes to the C++ side of the code
- Put all of the actual code in a single source file, because we don't
need a header there
- Made `WASMC` use `clang++` by default, because the `CXX` won't work in
many cases (GCC doesn't support WASM)
- Got rid of a completely unused variable in `sprintd` (`n`)
- Found a way to get rid of the ugliest line of `collatz_steps`
- Made `i` in the benchmark an `i32` (`uint_fast32_t`) instead of a
`uint_fast64_t`, because that's what it's used as
- Got rid of the extra defined struct `test_case` in the test, because
we only need it in one place and C++ is one of the languages that
provides that solution
- Actually we can use range-for in the test
- Removed the timing code from the test, because that's what a benchmark
is for

+++ test/benchmark.cc
@@ -1,9 +1,8 @@
-#include <collatz.hh>
-#include <stdint.h>
+#include <collatz.cc>
 
 int main() {
         volatile i32 j;
-        for(uint_fast64_t i = 0; i < 10000000; i++) {
+        for(i32 i = 0; i < 10000000; i++) {
                 j = collatz_steps(i);
         }
 }
